backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
fastapi-users[sqlalchemy]==13.0.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
python-multipart
